feat: action panel rework
