+++ src/server/CMakeLists.txt
@@ -181,6 +181,10 @@ set(SRCS
 
 	include/ui/action-pannel/action.hpp
 	src/ui/action-pannel/action.cpp
+	include/ui/action-pannel/action-panel-state.hpp
+	include/ui/action-pannel/action-panel-view.hpp
+	include/ui/action-pannel/action-list-view.hpp
+	src/ui/action-pannel/action-list-view.cpp
 
 	src/lib/crypto.cpp
